Renamed m2ts/captions-parser.js -> tools/cea708-parser.js
