+++ lib/m2ts/caption-stream.js
@@ -18,10 +18,10 @@
 // -----------------
 
 var Stream = require('../utils/stream');
-var USER_DATA_REGISTERED_ITU_T_T35 = require('./captions-parser').USER_DATA_REGISTERED_ITU_T_T35;
-var parseSei = require('./captions-parser').parseSei;
-var parseUserData = require('./captions-parser').parseUserData;
-var parseCaptionPackets = require('./captions-parser').parseCaptionPackets;
+var USER_DATA_REGISTERED_ITU_T_T35 = require('../tools/cea708-parser').USER_DATA_REGISTERED_ITU_T_T35;
+var parseSei = require('../tools/cea708-parser').parseSei;
+var parseUserData = require('../tools/cea708-parser').parseUserData;
+var parseCaptionPackets = require('../tools/cea708-parser').parseCaptionPackets;
 
 var CaptionStream = function() {
