+++ lib/mp4/captions-parser.js
@@ -8,7 +8,7 @@
  */
 'use strict';
 
-var discardEmulationPreventionBytes = require('../m2ts/captions-parser').discardEmulationPreventionBytes;
+var discardEmulationPreventionBytes = require('../tools/cea708-parser').discardEmulationPreventionBytes;
 var CaptionStream = require('../m2ts/caption-stream').CaptionStream;
 var mp4Probe = require('./probe');
 var findBox = mp4Probe.findBox;
